skip file

+++ removeEmptyDir.py
@@ -6,6 +6,11 @@ def removeEmptyDir(parentsDir):
     for item in os.scandir(parentsDir):
         if item.name == "ORGANIZED":
             continue
+
+        # skip file
+        if os.path.isfile(item):
+            continue
+
         shutil.rmtree(Path(item))
 
 def walkInDir():
